+++ projects/helper/sushi-server/adapterMapping.js
@@ -7,6 +7,7 @@ module.exports = {
         'astar-dapps-staking': '/astar-dapps-staking/api',
         'injective-orderbook': '/injective-orderbook/api',
         'lisa': '/lisa/api',
+        'chutes': '/chutes/api',
       },
       {
         'bifrost-staking': '/bifrost-staking/api',
